refactor(api): extract shared email thread matcher from inboundEmailService
